requirements.txt
streamlit
google-cloud-storage
google-cloud-firestore
ffmpeg-python
opencv-python-headless
torch
Pillow
git+https://github.com/openai/whisper.git
moviepy
streamlit-drawable-canvas
google-generativeai
